Datatypes optimization

+++ src/pandemia/components/health_model/default_health_model.py
@@ -8,7 +8,7 @@ import numpy as np
 from collections import defaultdict
 from numpy import genfromtxt
 
-from ctypes import c_void_p, c_double, c_int, cdll
+from ctypes import c_void_p, c_double, c_int32, cdll
 
 from ..health_model import HealthModel
 
@@ -52,7 +52,7 @@ class DefaultHealthModel(HealthModel):
         self.transmission.restype  = None
         self.infect.restype        = None
 
-        self.beta                             = np.array(config['beta'], dtype=np.float64)
+        self.beta                             = np.array(config['beta'], dtype=float)
 
         self.number_of_strains                = config['number_of_strains']
         self.num_initial_infections_by_region = config['num_initial_infections_by_region_by_strain']
@@ -126,98 +126,98 @@ class DefaultHealthModel(HealthModel):
                       self.number_of_strains,
                       self.number_of_strains,
                       self.max_preset_length_immunity,
-                      self.number_of_rho_immunity_outcomes), dtype=np.float64)
+                      self.number_of_rho_immunity_outcomes), dtype=float)
 
         self.preset_rho_immunity_failure_partitions =\
             np.zeros((self.number_of_presets,
                       self.number_of_rho_immunity_outcomes,
                       self.number_of_strains,
                       self.number_of_strains,
-                      self.max_preset_length_immunity), dtype=np.int64)
+                      self.max_preset_length_immunity), dtype=np.int32)
 
         self.preset_rho_immunity_failure_lengths =\
             np.zeros((self.number_of_presets,
                       self.number_of_rho_immunity_outcomes,
                       self.number_of_strains,
-                      self.number_of_strains), dtype=np.int64)
+                      self.number_of_strains), dtype=np.int32)
 
         self.preset_sigma_immunity_failure_values =\
             np.zeros((self.number_of_presets,
                       self.number_of_rho_immunity_outcomes,
                       self.number_of_strains,
                       self.number_of_strains,
-                      self.max_preset_length_immunity), dtype=np.float64)
+                      self.max_preset_length_immunity), dtype=float)
 
         self.preset_sigma_immunity_failure_partitions =\
             np.zeros((self.number_of_presets,
                       self.number_of_rho_immunity_outcomes,
                       self.number_of_strains,
                       self.number_of_strains,
-                      self.max_preset_length_immunity), dtype=np.int64)
+                      self.max_preset_length_immunity), dtype=np.int32)
 
         self.preset_sigma_immunity_failure_lengths =\
             np.zeros((self.number_of_presets,
                       self.number_of_rho_immunity_outcomes,
                       self.number_of_strains,
-                      self.number_of_strains), dtype=np.int64)
+                      self.number_of_strains), dtype=np.int32)
 
         self.preset_infectiousness_values =\
             np.zeros((self.number_of_presets,
                       self.number_of_rho_immunity_outcomes,
                       self.number_of_strains,
-                      self.max_preset_length_health), dtype=np.float64)
+                      self.max_preset_length_health), dtype=float)
 
         self.preset_infectiousness_partitions =\
             np.zeros((self.number_of_presets,
                       self.number_of_rho_immunity_outcomes,
                       self.number_of_strains,
-                      self.max_preset_length_health), dtype=np.int64)
+                      self.max_preset_length_health), dtype=np.int32)
 
         self.preset_infectiousness_lengths =\
             np.zeros((self.number_of_presets,
                       self.number_of_rho_immunity_outcomes,
-                      self.number_of_strains), dtype=np.int64)
+                      self.number_of_strains), dtype=np.int32)
 
         self.preset_disease_values =\
             np.zeros((self.number_of_presets,
                       self.number_of_rho_immunity_outcomes,
                       self.number_of_strains,
-                      self.max_preset_length_health), dtype=np.float64)
+                      self.max_preset_length_health), dtype=float)
 
         self.preset_disease_partitions =\
             np.zeros((self.number_of_presets,
                       self.number_of_rho_immunity_outcomes,
                       self.number_of_strains,
-                      self.max_preset_length_health), dtype=np.int64)
+                      self.max_preset_length_health), dtype=np.int32)
 
         self.preset_disease_lengths =\
             np.zeros((self.number_of_presets,
                       self.number_of_rho_immunity_outcomes,
-                      self.number_of_strains), dtype=np.int64)
+                      self.number_of_strains), dtype=np.int32)
 
         self.preset_strain_values =\
             np.zeros((self.number_of_presets,
                       self.number_of_rho_immunity_outcomes,
                       self.number_of_strains,
-                      self.max_preset_length_health), dtype=np.int64)
+                      self.max_preset_length_health), dtype=np.int32)
 
         self.preset_strain_partitions =\
             np.zeros((self.number_of_presets,
                       self.number_of_rho_immunity_outcomes,
                       self.number_of_strains,
-                      self.max_preset_length_health), dtype=np.int64)
+                      self.max_preset_length_health), dtype=np.int32)
 
         self.preset_strain_lengths =\
             np.zeros((self.number_of_presets,
                       self.number_of_rho_immunity_outcomes,
-                      self.number_of_strains), dtype=np.int64)
+                      self.number_of_strains), dtype=np.int32)
 
         # Create presets, thereby filling the above arrays, and validate them
         self._get_presets(self.clock.ticks_in_day)
         self._validate_presets()
 
         # Flatten arrays
-        self.mutation_matrix = np.asarray(self.mutation_matrix, dtype=np.float64).flatten()
+        self.mutation_matrix = np.asarray(self.mutation_matrix, dtype=float).flatten()
         self.preset_infectiousness_lengths =\
             self.preset_infectiousness_lengths.flatten()
         self.preset_infectiousness_values =\
@@ -259,52 +259,52 @@ class DefaultHealthModel(HealthModel):
         immunity_length                 = self.immunity_length
         max_preset_length_health        = self.max_preset_length_health
 
-        vector_region.health_age_group                 = np.zeros((number_of_agents), dtype=np.int64)
-        vector_region.location_transmission_multiplier = np.ones((number_of_locations), dtype=np.float64)
-        vector_region.infection_event                  = np.full((number_of_agents), -1, dtype=np.int64)
-        vector_region.presets                          = np.zeros((number_of_agents), dtype=np.int64)
-        vector_region.requesting_immunity_update       = np.zeros((number_of_agents), dtype=np.int64)
+        vector_region.health_age_group                 = np.zeros((number_of_agents), dtype=np.int32)
+        vector_region.location_transmission_multiplier = np.ones((number_of_locations), dtype=float)
+        vector_region.infection_event                  = np.full((number_of_agents), -1, dtype=np.int32)
+        vector_region.presets                          = np.zeros((number_of_agents), dtype=np.int32)
+        vector_region.requesting_immunity_update       = np.zeros((number_of_agents), dtype=np.int32)
 
         vector_region.current_rho_immunity_failure   = np.ones((number_of_agents,
                                                                 number_of_strains,
                                                                 number_of_rho_immunity_outcomes),
-                                                                dtype=np.float64)
+                                                                dtype=float)
         vector_region.current_sigma_immunity_failure = np.ones((number_of_agents,
                                                                 number_of_strains),
-                                                                dtype=np.float64)
+                                                                dtype=float)
 
-        vector_region.current_infectiousness = np.zeros((number_of_agents), dtype=np.float64)
-        vector_region.current_disease        = np.zeros((number_of_agents), dtype=np.float64)
-        vector_region.current_strain         = np.full((number_of_agents), -1, dtype=np.int64)
+        vector_region.current_infectiousness = np.zeros((number_of_agents), dtype=float)
+        vector_region.current_disease        = np.zeros((number_of_agents), dtype=float)
+        vector_region.current_strain         = np.full((number_of_agents), -1, dtype=np.int32)
 
         vector_region.rho_immunity_failure_values   = np.ones((number_of_agents,
                                                                number_of_strains,
                                                                immunity_length,
                                                                number_of_rho_immunity_outcomes),
-                                                               dtype=np.float64)
+                                                               dtype=float)
         vector_region.sigma_immunity_failure_values = np.ones((number_of_agents,
                                                                number_of_strains,
                                                                immunity_length),
-                                                               dtype=np.float64)
+                                                               dtype=float)
 
         vector_region.infectiousness_values     = np.zeros((number_of_agents,
-                                                            max_preset_length_health), dtype=np.float64)
+                                                            max_preset_length_health), dtype=float)
         vector_region.infectiousness_partitions = np.zeros((number_of_agents,
-                                                            max_preset_length_health), dtype=np.int64)
-        vector_region.infectiousness_lengths    = np.zeros((number_of_agents), dtype=np.int64)
-        vector_region.infectiousness_indexes    = np.ones((number_of_agents), dtype=np.int64)
+                                                            max_preset_length_health), dtype=np.int32)
+        vector_region.infectiousness_lengths    = np.zeros((number_of_agents), dtype=np.int32)
+        vector_region.infectiousness_indexes    = np.ones((number_of_agents), dtype=np.int32)
         vector_region.disease_values            = np.zeros((number_of_agents,
-                                                            max_preset_length_health), dtype=np.float64)
+                                                            max_preset_length_health), dtype=float)
         vector_region.disease_partitions        = np.zeros((number_of_agents,
-                                                            max_preset_length_health), dtype=np.int64)
-        vector_region.disease_lengths           = np.zeros((number_of_agents), dtype=np.int64)
-        vector_region.disease_indexes           = np.ones((number_of_agents), dtype=np.int64)
+                                                            max_preset_length_health), dtype=np.int32)
+        vector_region.disease_lengths           = np.zeros((number_of_agents), dtype=np.int32)
+        vector_region.disease_indexes           = np.ones((number_of_agents), dtype=np.int32)
         vector_region.strain_values             = np.full((number_of_agents,
-                                                           max_preset_length_health), -1, dtype=np.int64)
+                                                           max_preset_length_health), -1, dtype=np.int32)
         vector_region.strain_partitions         = np.zeros((number_of_agents,
-                                                            max_preset_length_health), dtype=np.int64)
-        vector_region.strain_lengths            = np.zeros((number_of_agents), dtype=np.int64)
-        vector_region.strain_indexes            = np.ones((number_of_agents), dtype=np.int64)
+                                                            max_preset_length_health), dtype=np.int32)
+        vector_region.strain_lengths            = np.zeros((number_of_agents), dtype=np.int32)
+        vector_region.strain_indexes            = np.ones((number_of_agents), dtype=np.int32)
 
         self._get_age_mixing_matrix(vector_region)
 
@@ -428,13 +428,13 @@ class DefaultHealthModel(HealthModel):
         """Updates agent health."""
 
         self.update_health(
-            c_int(t),
-            c_int(vector_region.number_of_agents),
-            c_int(self.number_of_strains),
-            c_int(self.immunity_length),
-            c_int(self.number_of_rho_immunity_outcomes),
-            c_int(self.max_preset_length_health),
-            c_int(self.immunity_period_ticks),
+            c_int32(t),
+            c_int32(vector_region.number_of_agents),
+            c_int32(self.number_of_strains),
+            c_int32(self.immunity_length),
+            c_int32(self.number_of_rho_immunity_outcomes),
+            c_int32(self.max_preset_length_health),
+            c_int32(self.immunity_period_ticks),
             c_void_p(vector_region.infectiousness_partitions.ctypes.data),
             c_void_p(vector_region.infectiousness_indexes.ctypes.data),
             c_void_p(vector_region.infectiousness_values.ctypes.data),
@@ -461,13 +461,13 @@ class DefaultHealthModel(HealthModel):
         """The transmission dynamics."""
 
         self.transmission(
-            c_int(self.number_of_strains),
-            c_int(vector_region.number_of_locations),
-            c_int(vector_region.number_of_agents),
-            c_int(vector_region.id),
-            c_int(self.sir_rescaling_int),
-            c_int(self.clock.ticks_in_day),
-            c_int(vector_region.number_of_subpopulations),
+            c_int32(self.number_of_strains),
+            c_int32(vector_region.number_of_locations),
+            c_int32(vector_region.number_of_agents),
+            c_int32(vector_region.id),
+            c_int32(self.sir_rescaling_int),
+            c_int32(self.clock.ticks_in_day),
+            c_int32(vector_region.number_of_subpopulations),
             c_void_p(self.beta.ctypes.data),
             c_void_p(vector_region.subpopulation_index.ctypes.data),
             c_void_p(vector_region.subpopulation_mixing_matrix.ctypes.data),
@@ -492,14 +492,14 @@ class DefaultHealthModel(HealthModel):
         """The infection system."""
 
         self.infect(
-            c_int(t),
-            c_int(vector_region.number_of_agents),
-            c_int(self.number_of_rho_immunity_outcomes),
-            c_int(self.number_of_strains),
-            c_int(self.max_preset_length_health),
-            c_int(self.max_preset_length_immunity),
-            c_int(self.immunity_length),
-            c_int(self.immunity_period_ticks),
+            c_int32(t),
+            c_int32(vector_region.number_of_agents),
+            c_int32(self.number_of_rho_immunity_outcomes),
+            c_int32(self.number_of_strains),
+            c_int32(self.max_preset_length_health),
+            c_int32(self.max_preset_length_immunity),
+            c_int32(self.immunity_length),
+            c_int32(self.immunity_period_ticks),
             c_void_p(vector_region.current_rho_immunity_failure.ctypes.data),
             c_void_p(vector_region.infection_event.ctypes.data),
             c_void_p(vector_region.infectiousness_partitions.ctypes.data),
@@ -721,7 +721,7 @@ class DefaultHealthModel(HealthModel):
                 if vector_region.name + ".csv" in files:
                     mat = genfromtxt(self.age_mixing_matrices_directory +\
                                      vector_region.name + ".csv", skip_header = 1,
-                                     delimiter=',', dtype=np.float64)
+                                     delimiter=',', dtype=float)
                     assert mat.shape[0] == mat.shape[1]
                     vector_region.number_of_subpopulations = mat.shape[0]
 
@@ -738,7 +738,7 @@ class DefaultHealthModel(HealthModel):
                     mat = mat * row_sums[:, np.newaxis]
 
                     # Calculate share of population in each age mixing group
-                    share_in_age_group = np.zeros((mat.shape[0]), dtype=np.float64)
+                    share_in_age_group = np.zeros((mat.shape[0]), dtype=float)
                     for n in range(vector_region.number_of_agents):
                         age_group = min(vector_region.age[n] // self.age_group_interval,
                                         mat.shape[0] - 1)
@@ -749,21 +749,21 @@ class DefaultHealthModel(HealthModel):
                     # share_in_age_group is equal to 1.0, that is, the same for each region
                     dim = vector_region.number_of_subpopulations
                     epsilon = (1.0 - np.inner(mat.sum(axis=1), share_in_age_group)) / dim
-                    mat += np.full((dim, dim), epsilon, dtype=np.float64)
+                    mat += np.full((dim, dim), epsilon, dtype=float)
                     vector_region.subpopulation_mixing_matrix = mat
 
                 else:
-                    vector_region.subpopulation_mixing_matrix = np.ones((1, 1), dtype=np.float64)
+                    vector_region.subpopulation_mixing_matrix = np.ones((1, 1), dtype=float)
                     vector_region.number_of_subpopulations = 1
-            vector_region.age_group = np.zeros((vector_region.number_of_agents), dtype=np.int64)
+            vector_region.age_group = np.zeros((vector_region.number_of_agents), dtype=np.int32)
             for n in range(vector_region.number_of_agents):
                 age_group = min(vector_region.age[n] // self.age_group_interval,
                                 vector_region.number_of_subpopulations - 1)
                 vector_region.subpopulation_index[n] = age_group
         else:
-            vector_region.subpopulation_mixing_matrix = np.ones((1, 1), dtype=np.float64)
+            vector_region.subpopulation_mixing_matrix = np.ones((1, 1), dtype=float)
             vector_region.subpopulation_index =\
-                np.zeros((vector_region.number_of_agents), dtype=np.int64)
+                np.zeros((vector_region.number_of_agents), dtype=np.int32)
             vector_region.number_of_subpopulations = 1
 
     def _generate_presets_and_weights(self, sir_beta, sir_gamma_inverse,
